Codechange: split off the settings saveload code from the main settings handling logic
